+++ include/shisen/camera.hpp
@@ -5,6 +5,7 @@
 #include <rclcpp/rclcpp.hpp>
 #include <sensor_msgs/msg/compressed_image.hpp>
 
+#include <map>
 #include <memory>
 #include <string>
 
@@ -15,13 +16,32 @@ namespace shisen
   public:
 
     Camera(std::string node_name);
+    ~Camera();
 
     void open(std::string device_name);
     void open();
 
+    void close();
+
   private:
 
+    const std::map<std::string, int> property_names = {
+      { "brightness", cv::CAP_PROP_BRIGHTNESS },
+      { "contrast", cv::CAP_PROP_CONTRAST },
+      { "saturation", cv::CAP_PROP_SATURATION },
+      { "gain", cv::CAP_PROP_GAIN },
+      { "exposure", cv::CAP_PROP_EXPOSURE },
+      { "temperature", cv::CAP_PROP_TEMPERATURE }
+    };
+
     cv::VideoCapture video_capture;
+
+    std::shared_ptr<OnSetParametersCallbackHandle> on_set_parameter_handler;
+
+    std::shared_ptr<rclcpp::AsyncParametersClient> parameter_client;
+
+    std::shared_ptr<rclcpp::Subscription<rcl_interfaces::msg::ParameterEvent>>
+      parameter_event_subscription;
 
     std::shared_ptr<rclcpp::Publisher<sensor_msgs::msg::CompressedImage>>
       compressed_image_publisher;